Skipping user deletion if service is read-only.
Adding isReadOnly() to user service API, and using that to skip user deletion
when an owner is cleaned up.

+++ proxy/src/main/java/org/fedoraproject/candlepin/resource/OwnerResource.java
@@ -236,8 +236,10 @@ public class OwnerResource {
 
     private void cleanupAndDelete(Owner owner, boolean revokeCerts) {
         log.info("Cleaning up owner: " + owner);
-        for (User u : userService.listByOwner(owner)) {
-            userService.deleteUser(u);
+        if (!userService.isReadyOnly()) {
+            for (User u : userService.listByOwner(owner)) {
+                userService.deleteUser(u);
+            }
         }
         for (Consumer c : consumerCurator.listByOwner(owner)) {
             log.info("Deleting consumer: " + c);
